+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versions since `1.7.0` only track ABI breaks and not API breaks.
 ## [Unreleased]
 ### Added
 - Libtcod can now be built without zlib by passing the `-DLIBTCOD_ZLIB=disable` flag to CMake or by setting `CMAKE_DISABLE_FIND_PACKAGE_ZLIB=1`.
+- Added context screen capture functions to take screenshots into memory instead of a file.
 
 ### Changed
 - The OpenGL2 renderer now defaults to the nearest filter unless `SDL_HINT_RENDER_SCALE_QUALITY` is set otherwise.